+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 ## 12.x.x (unreleased)
 
 ### Improvements
-- Client: Preferences Registry
 
 ### Bugfixes
 - 
@@ -13,6 +12,14 @@ You should also include the user name that made the change.
 ## 12.x.x (unreleased)
 
 ### Improvements
+
+### Bugfixes
+- Client: can not show some setting pages @syuilo
+
+## 12.118.0 (2022/08/07)
+
+### Improvements
+- Client: 設定のバックアップ/リストア機能
 - Client: Add vi-VN language support
 - Client: Add unix time widget @syuilo